Feat: 게시글 헤더, 바디 컴포넌트 분리
