+++ src/components/common/PostContent.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="post-content">
+        <slot></slot>
+    </div>
+</template>
