+++ src/views/Group/page/StudyBoardDetailPage.vue
@@ -1,0 +1,111 @@
+<template>
+    <div class="study-board-detail-page">
+        <!-- 상단 네비게이션 -->
+        <Navigation />
+        <!-- 사이드 바-->
+        <GroupSideBar />
+        <main class="main">
+            <div class="main-content">
+                <div class="post-container" v-if="loading">Loading...</div>
+                <div class="post-container" v-else>
+                    <Title>{{title}}</Title>
+                    <PostHeader :data="headerData" :isNotice="false" @deletePost="deletePost"></PostHeader>
+                    <PostBody :data="bodyData" :isPerm="true"></PostBody>
+                    <DeleteModal :isVisible="modalVisibility" @confirm="confirm" @cancel="cancel">해당 게시글을 삭제하시겠습니까?</DeleteModal>
+                </div>
+            </div>
+        </main>
+    </div>
+</template>
+
+<script setup>
+    import Navigation from '@/components/layouts/Navigation.vue';
+    import GroupSideBar from '@/components/layouts/GroupSideBar.vue';
+    import Title from '@/components/common/Title.vue';
+    import PostHeader from '@/components/layouts/PostHeader.vue';
+    import PostBody from '@/components/layouts/PostBody.vue';
+    import DeleteModal from '@/components/common/DeleteModal.vue';
+    import { ref, onMounted } from 'vue';
+    import axios from 'axios';
+
+    const props = defineProps({
+        id: {
+            type: String,
+            required: true
+        }
+    })
+
+    const item = ref({});
+    const loading = ref(true);
+    const modalVisibility = ref(false);
+
+    const title = ref(null);
+    const nickname = ref(null);
+    const likes = ref(0);
+
+    const startDate = ref(null);
+    const endDate = ref(null);
+    const content = ref(null);
+
+    const headerData = {
+        nickname: nickname,
+        likes: likes,
+    }
+
+    const bodyData = {
+        startDate: null,
+        endDate: null,
+        content: content,
+    }
+
+    const fetchData = async () => {
+      try {
+        let response; // response 변수를 미리 선언
+
+        response = await axios.get(`/api/study-group/boards/${props.id}`);
+        item.value = response.data.data;
+
+        title.value = item.value.title;
+        nickname.value = item.value.nickname;
+        likes.value = item.value.likes;
+
+        if (typeof item.value.schedule_start_time !== 'undefined') {
+            startDate.value = item.value.schedule_start_time;
+            endDate.value = item.value.schedule_end_time;
+        } else if (typeof item.value.recruitment_start_time !== 'undefined') {
+            startDate.value = item.value.recruitment_start_time;
+            endDate.value = item.value.recruitment_end_time;
+        }
+        content.value = item.value.content;
+
+      } catch (error) {
+        console.error(error);
+      } finally {
+        loading.value = false;
+      }
+    }
+
+    const deletePost = () => {
+        modalVisibility.value = true;
+    };
+
+    const confirm = () => {
+        modalVisibility.value = false;
+    }
+
+    const cancel = () => {
+        modalVisibility.value = false;
+    }
+
+    onMounted(() => {
+        fetchData();
+    })
+</script>
+
+<style scoped>
+    .post-container {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+    }
+</style>
